Let agent.Config control intercepts in the traffic-agent
The traffic-agent is rewritten in this commit so that it no longer uses
environment variables for its configuration. Instead it uses the new
`agent.Config` that it gets mounted from a `ConfigMap` entry.

This change makes the agent multi-port aware, i.e. it can have multiple
simultaneous intercepts on different container ports.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>