Add multi-port to CLI and make the user daemon use the agents ConfigMap
Moves the bulk of `CanIntercept` to the cluster-side traffic-manager
in the form of a `PrepareIntercept` method. This method is responsible
for creating the initial agent `ConfigMap` in a namespace and to add
the `create` request into that is picked up by the watcher and results
in a generated entry.

All `Uninstall` functionality is now done using the helm chart and the
agent `ConfigMap`.

The `telepresesence intercept --port` flag can now be repeated and
multiple `ServicePortIdentifier` / `TargetPort` pairs can be propagated
to the agent.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>